Safari-611.1.10/LayoutTests/media/W3C/audio/error/error_null.html

<!doctype html>
<html>
 <head>
  <title>audio.error - null if no source</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-error">spec reference</a></p>
  <audio id="a">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("a").error,
  null,
  "audioElement.error property is null if no source");
});
  </script>
 </body>
</html>
